+++ packages/core/src/components/fields/ObjectField.tsx
@@ -67,7 +67,11 @@ class ObjectField<
    * @returns - The onPropertyChange callback for the `name` property
    */
   onPropertyChange = (name: string, addedByAdditionalProperties = false) => {
-    return (value: T, newErrorSchema?: ErrorSchema<T>, id?: string) => {
+    return (
+      value: T | undefined,
+      newErrorSchema?: ErrorSchema<T>,
+      id?: string
+    ) => {
       const { formData, onChange, errorSchema } = this.props;
       if (value === undefined && addedByAdditionalProperties) {
         // Don't set value = undefined for fields added by
@@ -79,7 +83,7 @@ class ObjectField<
         // set empty values to the empty string.
         value = "" as unknown as T;
       }
-      const newFormData = { ...formData, [name]: value };
+      const newFormData = { ...formData, [name]: value } as unknown as T;
       onChange(
         newFormData,
         errorSchema &&
@@ -102,7 +106,7 @@ class ObjectField<
     return (event: DragEvent) => {
       event.preventDefault();
       const { onChange, formData } = this.props;
-      const copiedFormData = { ...formData };
+      const copiedFormData = { ...formData } as T;
       unset(copiedFormData, key);
       onChange(copiedFormData);
     };
@@ -112,10 +116,10 @@ class ObjectField<
    * that is already not assigned is found.
    *
    * @param preferredKey - The preferred name of a new key
-   * @param formData - The form data in which to check if the desired key already exists
+   * @param [formData] - The form data in which to check if the desired key already exists
    * @returns - The name of the next available key from `preferredKey`
    */
-  getAvailableKey = (preferredKey: string, formData: T) => {
+  getAvailableKey = (preferredKey: string, formData?: T) => {
     const { uiSchema } = this.props;
     const { duplicateKeySuffixSeparator = "-" } = getUiOptions<T, S, F>(
       uiSchema
@@ -200,7 +204,7 @@ class ObjectField<
       return;
     }
     const { formData, onChange, registry } = this.props;
-    const newFormData = { ...formData };
+    const newFormData = { ...formData } as T;
 
     let type: RJSFSchema["type"] = undefined;
     if (isObject(schema.additionalProperties)) {